repo level for forum in process

+++ internal/forum/delivery/models/createforum.go
@@ -66,8 +66,8 @@ type ForumCreateResponse struct {
 	Title   string `json:"title"`
 	User    string `json:"user"`
 	Slug    string `json:"slug"`
-	Posts   uint32 `json:"posts"`
-	Threads uint32 `json:"threads"`
+	Posts   int64  `json:"posts"`
+	Threads int64  `json:"threads"`
 }
 
 func NewForumCreateResponse(forum *models.Forum) *ForumCreateResponse {
