+++ internal/models/thread.go
@@ -1,12 +1,12 @@
 package models
 
 type Thread struct {
-	ID      uint32
+	ID      int64
 	Title   string
 	Author  string
 	Forum   string
 	Slug    string
 	Message string
 	Created string
-	Votes   int32
+	Votes   int64
 }
