+++ internal/service/delivery/models/getstatus.go
@@ -5,10 +5,10 @@ import "db-performance-project/internal/models"
 //go:generate easyjson -disallow_unknown_fields getstatus.go
 
 type ServiceGetStatusResponse struct {
-	User   uint32 `json:"user"`
-	Forum  uint32 `json:"forum"`
-	Thread uint32 `json:"thread"`
-	Post   uint32 `json:"post"`
+	User   int64 `json:"user"`
+	Forum  int64 `json:"forum"`
+	Thread int64 `json:"thread"`
+	Post   int64 `json:"post"`
 }
 
 func NewServiceGetStatusResponse(service *models.StatusService) *ServiceGetStatusResponse {
