+++ internal/thread/delivery/models/createposts_easyjson.go
@@ -169,9 +169,9 @@ func easyjson16fc04efDecodeDbPerformanceProjectInternalThreadDeliveryModels2(in 
 		}
 		switch key {
 		case "id":
-			out.ID = uint32(in.Uint32())
+			out.ID = int64(in.Int64())
 		case "parent":
-			out.Parent = uint32(in.Uint32())
+			out.Parent = int64(in.Int64())
 		case "author":
 			out.Author = string(in.String())
 		case "message":
@@ -181,7 +181,7 @@ func easyjson16fc04efDecodeDbPerformanceProjectInternalThreadDeliveryModels2(in 
 		case "forum":
 			out.Forum = string(in.String())
 		case "thread":
-			out.Thread = uint32(in.Uint32())
+			out.Thread = int64(in.Int64())
 		case "created":
 			out.Created = string(in.String())
 		default:
@@ -205,12 +205,12 @@ func easyjson16fc04efEncodeDbPerformanceProjectInternalThreadDeliveryModels2(out
 	{
 		const prefix string = ",\"id\":"
 		out.RawString(prefix[1:])
-		out.Uint32(uint32(in.ID))
+		out.Int64(int64(in.ID))
 	}
 	{
 		const prefix string = ",\"parent\":"
 		out.RawString(prefix)
-		out.Uint32(uint32(in.Parent))
+		out.Int64(int64(in.Parent))
 	}
 	{
 		const prefix string = ",\"author\":"
@@ -235,7 +235,7 @@ func easyjson16fc04efEncodeDbPerformanceProjectInternalThreadDeliveryModels2(out
 	{
 		const prefix string = ",\"thread\":"
 		out.RawString(prefix)
-		out.Uint32(uint32(in.Thread))
+		out.Int64(int64(in.Thread))
 	}
 	{
 		const prefix string = ",\"created\":"
@@ -288,7 +288,7 @@ func easyjson16fc04efDecodeDbPerformanceProjectInternalThreadDeliveryModels3(in 
 		}
 		switch key {
 		case "parent":
-			out.Parent = uint32(in.Uint32())
+			out.Parent = int64(in.Int64())
 		case "author":
 			out.Author = string(in.String())
 		case "message":
@@ -314,7 +314,7 @@ func easyjson16fc04efEncodeDbPerformanceProjectInternalThreadDeliveryModels3(out
 	{
 		const prefix string = ",\"parent\":"
 		out.RawString(prefix[1:])
-		out.Uint32(uint32(in.Parent))
+		out.Int64(int64(in.Parent))
 	}
 	{
 		const prefix string = ",\"author\":"
