+++ internal/thread/delivery/models/createthread_easyjson.go
@@ -37,7 +37,7 @@ func easyjson9d4b6d10DecodeDbPerformanceProjectInternalThreadDeliveryModels(in *
 		}
 		switch key {
 		case "id":
-			out.ID = uint32(in.Uint32())
+			out.ID = int64(in.Int64())
 		case "title":
 			out.Title = string(in.String())
 		case "author":
@@ -51,7 +51,7 @@ func easyjson9d4b6d10DecodeDbPerformanceProjectInternalThreadDeliveryModels(in *
 		case "created":
 			out.Created = string(in.String())
 		case "votes":
-			out.Votes = int32(in.Int32())
+			out.Votes = int64(in.Int64())
 		default:
 			in.AddError(&jlexer.LexerError{
 				Offset: in.GetPos(),
@@ -73,7 +73,7 @@ func easyjson9d4b6d10EncodeDbPerformanceProjectInternalThreadDeliveryModels(out 
 	{
 		const prefix string = ",\"id\":"
 		out.RawString(prefix[1:])
-		out.Uint32(uint32(in.ID))
+		out.Int64(int64(in.ID))
 	}
 	{
 		const prefix string = ",\"title\":"
@@ -108,7 +108,7 @@ func easyjson9d4b6d10EncodeDbPerformanceProjectInternalThreadDeliveryModels(out 
 	{
 		const prefix string = ",\"votes\":"
 		out.RawString(prefix)
-		out.Int32(int32(in.Votes))
+		out.Int64(int64(in.Votes))
 	}
 	out.RawByte('}')
 }
