+++ internal/thread/delivery/models/updatedetailsthread.go
@@ -64,7 +64,7 @@ func (req *ThreadUpdateDetailsRequest) GetThread() *models.Thread {
 	id, err := strconv.Atoi(req.SlugOrID)
 	if err != nil {
 		return &models.Thread{
-			ID: uint32(id),
+			ID: int64(id),
 		}
 	}
 
@@ -75,14 +75,14 @@ func (req *ThreadUpdateDetailsRequest) GetThread() *models.Thread {
 
 //easyjson:json
 type ThreadUpdateDetailsResponse struct {
-	ID      uint32 `json:"id"`
+	ID      int64  `json:"id"`
 	Title   string `json:"title"`
 	Author  string `json:"author"`
 	Forum   string `json:"forum"`
 	Slug    string `json:"slug"`
 	Message string `json:"message"`
 	Created string `json:"created"`
-	Votes   int32  `json:"votes"`
+	Votes   int64  `json:"votes"`
 }
 
 func NewThreadUpdateDetailsResponse(thread *models.Thread) *ThreadUpdateDetailsResponse {
